refactor: extract vial related code in storage
