+++ rmk/src/host/via/vial.rs
@@ -15,9 +15,8 @@ use crate::keymap::KeyMap;
 use crate::morse::{DOUBLE_TAP, HOLD, HOLD_AFTER_TAP, TAP};
 #[cfg(feature = "storage")]
 use crate::{
-    COMBO_MAX_LENGTH,
-    channel::FLASH_CHANNEL,
-    storage::{ComboData, FlashOperationMessage},
+    COMBO_MAX_LENGTH, channel::FLASH_CHANNEL, host::storage_types::ComboData, host::via::storage::VialData,
+    storage::FlashOperationMessage,
 };
 use crate::{COMBO_MAX_NUM, MORSE_MAX_NUM};
 
@@ -319,7 +318,10 @@ pub(crate) async fn process_vial<
                             {
                                 // Save to storage
                                 FLASH_CHANNEL
-                                    .send(FlashOperationMessage::WriteMorse(morse_idx as u8, morse.clone()))
+                                    .send(FlashOperationMessage::VialMessage(VialData::Morse(
+                                        morse_idx as u8,
+                                        morse.clone(),
+                                    )))
                                     .await;
                             }
                         }
@@ -389,11 +391,11 @@ pub(crate) async fn process_vial<
                     };
                     #[cfg(feature = "storage")]
                     FLASH_CHANNEL
-                        .send(FlashOperationMessage::WriteCombo(ComboData {
+                        .send(FlashOperationMessage::VialMessage(VialData::Combo(ComboData {
                             idx: real_idx,
                             actions,
                             output,
-                        }))
+                        })))
                         .await;
                 }
                 VialDynamic::DynamicVialKeyOverrideGet => {
@@ -469,12 +471,13 @@ pub(crate) async fn process_vial<
             // Save the encoder action to the storage after the RefCell is released
             if let Some(encoder) = _encoder {
                 // Save the encoder action to the storage
+                use crate::host::storage_types::EncoderConfig;
                 FLASH_CHANNEL
-                    .send(FlashOperationMessage::EncoderKey {
+                    .send(FlashOperationMessage::VialMessage(VialData::Encoder(EncoderConfig {
                         idx: index,
                         layer,
                         action: encoder,
-                    })
+                    })))
                     .await;
             }
         }
